refactor(dataset): restructure services into application layer
